Refactor logging in agentchat
