+++ python/packages/autogen-agentchat/src/autogen_agentchat/__init__.py
@@ -1,0 +1,2 @@
+TRACE_LOGGER_NAME = "autogen_agentchat"
+EVENT_LOGGER_NAME = "autogen_agentchat.events"
