+++ python/packages/autogen-agentchat/src/autogen_agentchat/teams/__init__.py
@@ -1,13 +1,8 @@
-from ._logging import EVENT_LOGGER_NAME, TRACE_LOGGER_NAME, ConsoleLogHandler, FileLogHandler
-from ._termination import MaxMessageTermination, StopMessageTermination, TerminationCondition, TextMentionTermination
 from ._group_chat._round_robin_group_chat import RoundRobinGroupChat
 from ._group_chat._selector_group_chat import SelectorGroupChat
+from ._termination import MaxMessageTermination, StopMessageTermination, TerminationCondition, TextMentionTermination
 
 __all__ = [
-    "TRACE_LOGGER_NAME",
-    "EVENT_LOGGER_NAME",
-    "ConsoleLogHandler",
-    "FileLogHandler",
     "TerminationCondition",
     "MaxMessageTermination",
     "TextMentionTermination",
