+++ python/packages/autogen-agentchat/src/autogen_agentchat/teams/_base_team.py
@@ -1,15 +1,8 @@
-import logging
 from dataclasses import dataclass
 from typing import List, Protocol
 
 from ..agents import ChatMessage
-from ._logging import EVENT_LOGGER_NAME, ConsoleLogHandler
 from ._termination import TerminationCondition
-
-logger = logging.getLogger(EVENT_LOGGER_NAME)
-logger.setLevel(logging.INFO)
-console_handler = ConsoleLogHandler()
-logger.addHandler(console_handler)
 
 
 @dataclass
